requirements.txt
names_generator==0.2.0
python-hcl2==7.2
lark-parser==0.12.0
dotenv==0.9.9
psycopg2==2.9.10
